Fix references to IDataElement, etc. after move
