+++ org.eclipse.ice.data/pom.xml
@@ -5,7 +5,7 @@
 
 	<groupId>org.eclipse.ice</groupId>
 	<artifactId>org.eclipse.ice.data</artifactId>
-	<version>3.0-SNAPSHOT</version>
+	<version>3.0.0-SNAPSHOT</version>
 	<packaging>jar</packaging>
 
 	<name>org.eclipse.ice.data</name>
