+++ org.eclipse.ice.dev/org.eclipse.ice.dev.annotations/pom.xml
@@ -100,7 +100,7 @@
 			<groupId>org.eclipse.ice</groupId>
 			<artifactId>org.eclipse.ice.data</artifactId>
 			<version>3.0.0-SNAPSHOT</version>
-			<scope>provided</scope>
+			<scope>compile</scope>
 		</dependency>
 	</dependencies>
