+++ org.eclipse.ice.dev/org.eclipse.ice.dev.annotations/src/main/java/org/eclipse/ice/dev/annotations/processors/DataElementSpec.java
@@ -12,7 +12,6 @@
 package org.eclipse.ice.dev.annotations.processors;
 
 import java.util.Arrays;
-
 import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
