+++ org.eclipse.ice.dev/org.eclipse.ice.dev.annotations/src/main/resources/templates/DataElement.vm
@@ -22,6 +22,8 @@ import lombok.NoArgsConstructor;
 import lombok.NonNull;
 import lombok.Setter;
 
+import org.eclipse.ice.data.JavascriptValidator;
+
 /**
  * This is an implementation of $interface that satisfies the dependencies of
  * the @DataElement Annotation and was auto-generated by the ICE Framework.
