+++ org.eclipse.ice.dev/org.eclipse.ice.dev.annotations/src/test/resources/patterns/DataElement/Defaults.java
@@ -1,2 +1,2 @@
-import org.eclipse.ice.dev.annotations.IDataElement;
+import org.eclipse.ice.data.IDataElement;
 public interface Test extends IDataElement<Test> {}
