+++ org.eclipse.ice.dev/org.eclipse.ice.dev.annotations/src/test/resources/patterns/DataElement/Setter.java
@@ -1,4 +1,4 @@
-import org.eclipse.ice.dev.annotations.IDataElement;
+import org.eclipse.ice.data.IDataElement;
 public interface Test extends IDataElement<Test> {
 	public int getTest();
 }
